+++ neuron/src/app/Neuron/Web/Generate.hs
@@ -1,3 +1,4 @@
+{-# LANGUAGE BangPatterns #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE LambdaCase #-}
 {-# LANGUAGE OverloadedStrings #-}
@@ -13,18 +14,17 @@
 module Neuron.Web.Generate
   ( generateSite,
     loadZettelkasten,
-    loadZettelkastenGraph,
   )
 where
 
 import Control.Monad.Writer.Strict (runWriter, tell)
 import qualified Data.Map.Strict as Map
-import Data.Tagged (untag)
-import qualified Data.Text as T
 import Development.Shake (Action, need)
 import Neuron.Config.Type (Config)
-import qualified Neuron.Config.Type as C
-import Neuron.Version (neuronVersion, olderThan)
+import qualified Neuron.Config.Type as Config
+import Neuron.Plugin (PluginRegistry)
+import qualified Neuron.Plugin as Plugin
+import Neuron.Version (neuronVersion)
 import qualified Neuron.Web.Cache as Cache
 import Neuron.Web.Cache.Type (NeuronCache, _neuronCache_graph)
 import qualified Neuron.Web.Cache.Type as Cache
@@ -32,18 +32,25 @@ import Neuron.Web.Generate.Route ()
 import qualified Neuron.Web.Route as Z
 import qualified Neuron.Zettelkasten.Graph.Build as G
 import Neuron.Zettelkasten.Graph.Type (ZettelGraph, stripSurroundingContext)
-import Neuron.Zettelkasten.ID (ZettelID, getZettelID, unZettelID)
-import Neuron.Zettelkasten.Query.Error (showQueryResultError)
+import Neuron.Zettelkasten.ID (ZettelID (..))
+import qualified Neuron.Zettelkasten.Resolver as R
 import Neuron.Zettelkasten.Zettel
   ( ZettelC,
-    ZettelError (..),
     ZettelT (zettelSlug),
     sansContent,
   )
-import Neuron.Zettelkasten.Zettel.Parser (extractQueriesWithContext, parseZettels)
-import Relude
-import Rib.Shake (forEvery, ribInputDir)
+import Neuron.Zettelkasten.Zettel.Error
+  ( ZettelError (ZettelError_AmbiguousID),
+    zettelErrorList,
+  )
+import Relude hiding (traceShowId)
+import Rib.Shake (ribInputDir)
+import System.Directory (withCurrentDirectory)
+import qualified System.Directory.Contents as DC
+import System.Environment (lookupEnv)
 import System.FilePath ((</>))
+import System.TimeIt (timeItT)
+import Text.Printf (printf)
 
 -- | Generate the Zettelkasten site
 generateSite ::
@@ -51,122 +58,102 @@ generateSite ::
   (forall a. NeuronCache -> Z.Route a -> a -> Action ()) ->
   Action ZettelGraph
 generateSite config writeHtmlRoute' = do
-  when (olderThan $ C.minVersion config) $ do
-    fail $
-      toString $
-        "Require neuron mininum version "
-          <> C.minVersion config
-          <> ", but your neuron version is "
-          <> neuronVersion
-  (cache@Cache.NeuronCache {..}, zettelContents) <- loadZettelkasten config
+  reportPerf <- liftIO $ isJust <$> lookupEnv "NEURON_PERF"
+  (buildDur, (!cache@Cache.NeuronCache {..}, !zettelContents)) <- timeItT $ loadZettelkasten config
+  when reportPerf $ liftIO $ putStrLn $ printf "Took %.2fs to build the graph" buildDur
   let writeHtmlRoute :: forall a. a -> Z.Route a -> Action ()
       writeHtmlRoute v r = writeHtmlRoute' cache r v
-  -- Generate HTML for every zettel
-  forM_ zettelContents $ \val@(sansContent -> z) ->
-    writeHtmlRoute val $ Z.Route_Zettel (zettelSlug z)
-  -- Generate search page
-  writeHtmlRoute () $ Z.Route_Impulse Nothing
+  (writeDur, ()) <- timeItT $ do
+    -- Generate HTML for every zettel
+    forM_ zettelContents $ \val@(zettelSlug . sansContent -> slug) ->
+      writeHtmlRoute val $ Z.Route_Zettel slug
+    -- Generate search page
+    writeHtmlRoute () $ Z.Route_Impulse Nothing
   -- Report all errors
+  -- TODO: Report only new errors in this run, to avoid spamming the terminal.
   forM_ (Map.toList _neuronCache_errors) $ \(zid, err) -> do
-    reportError zid $
-      case err of
-        ZettelError_ParseError (untag . snd -> parseErr) ->
-          parseErr :| []
-        ZettelError_QueryResultErrors queryErrs ->
-          showQueryResultError <$> snd queryErrs
-        ZettelError_AmbiguousID filePaths ->
-          ("Multiple zettels have the same ID: " <> T.intercalate ", " (toText <$> toList filePaths))
-            :| []
-        ZettelError_AmbiguousSlug slug ->
-          "Slug '" <> slug <> "' is already used by another zettel" :| []
+    reportError zid $ zettelErrorList err
+  when reportPerf $ liftIO $ putStrLn $ printf "Took %.2fs to generate html" writeDur
   pure _neuronCache_graph
-
--- | Report an error in the terminal
-reportError :: MonadIO m => ZettelID -> NonEmpty Text -> m ()
-reportError zid errors = do
-  putTextLn $ "E " <> unZettelID zid
-  forM_ errors $ \err ->
-    putText $ "  - " <> indentAllButFirstLine 4 err
   where
-    indentAllButFirstLine :: Int -> Text -> Text
-    indentAllButFirstLine n = unlines . go . lines
+    -- Report an error in the terminal
+    reportError :: MonadIO m => ZettelID -> NonEmpty Text -> m ()
+    reportError zid errors = do
+      putTextLn $ "E " <> unZettelID zid
+      forM_ errors $ \err ->
+        putText $ "  - " <> indentAllButFirstLine 4 err
       where
-        go [] = []
-        go [x] = [x]
-        go (x : xs) =
-          x : fmap (toText . (replicate n ' ' <>) . toString) xs
-
--- | Like `loadZettelkasten` but without the content
---
--- Also allows retrieving the cached data for faster execution.
-loadZettelkastenGraph ::
-  Config ->
-  Action NeuronCache
-loadZettelkastenGraph =
-  fmap fst . loadZettelkasten
+        indentAllButFirstLine :: Int -> Text -> Text
+        indentAllButFirstLine n = unlines . go . lines
+          where
+            go [] = []
+            go [x] = [x]
+            go (x : xs) =
+              x : fmap (toText . (replicate n ' ' <>) . toString) xs
 
 loadZettelkasten :: Config -> Action (NeuronCache, [ZettelC])
 loadZettelkasten config = do
-  let pat = bool "*.md" "**/*.md" $ C.recurseDir config
-  files <- forEvery [pat] pure
-  ((g, zs), errs) <- loadZettelkastenFrom files
+  let plugins = Config.getPlugins config
+  liftIO $ putStrLn $ "Plugins enabled: " <> show (Map.keys plugins)
+  ((g, zs), errs) <- loadZettelkastenFromFiles plugins =<< locateZettelFiles plugins
   let cache = Cache.NeuronCache g errs config neuronVersion
       cacheSmall = cache {_neuronCache_graph = stripSurroundingContext g}
   Cache.updateCache cacheSmall
   pure (cache, zs)
 
+locateZettelFiles :: PluginRegistry -> Action (DC.DirTree FilePath)
+locateZettelFiles plugins = do
+  -- Run with notes dir as PWD, so that DirTree uses relative paths throughout.
+  d <- ribInputDir
+  liftIO $
+    withCurrentDirectory d $
+      -- Building with "." as root directory ensures that the returned tree
+      -- structure consistently uses the "./" prefix for all paths. This
+      -- assumption then holds elsewhere in neuron.
+      DC.buildDirTree "." >>= \case
+        Just t -> do
+          Plugin.filterSources plugins t >>= \case
+            Nothing ->
+              fail "No source files to process"
+            Just tF ->
+              pure tF
+        Nothing ->
+          fail "No sources"
+
 -- | Load the Zettelkasten from disk, using the given list of zettel files
-loadZettelkastenFrom ::
-  [FilePath] ->
+loadZettelkastenFromFiles ::
+  PluginRegistry ->
+  DC.DirTree FilePath ->
   Action
     ( ( ZettelGraph,
         [ZettelC]
       ),
       Map ZettelID ZettelError
     )
-loadZettelkastenFrom files = do
-  zidRefs <- resolveZidRefs files
+loadZettelkastenFromFiles plugins fileTree = do
+  let total = getSum @Int $ foldMap (const $ Sum 1) fileTree
+  -- TODO: Would be nice to show a progressbar here
+  -- liftIO $ DC.printDirTree fileTree
+  liftIO $ putStrLn $ "Loading directory tree (" <> show total <> " files) ..."
+  zidRefs <-
+    fmap snd $
+      flip runStateT Map.empty $ do
+        flip R.resolveZidRefsFromDirTree fileTree $ \relPath -> do
+          -- NOTE: This is the only place where Shake is being used (for
+          -- posterity)
+          absPath <- fmap (</> relPath) ribInputDir
+          need [absPath]
+          decodeUtf8With lenientDecode <$> readFileBS absPath
+        Plugin.afterZettelRead plugins fileTree
   pure $
     runWriter $ do
       filesWithContent <-
-        flip Map.traverseMaybeWithKey zidRefs $ \zid zidRef ->
-          case zidRef of
-            ZIDRef_Ambiguous fps -> do
-              tell $ one (zid, ZettelError_AmbiguousID fps)
-              pure Nothing
-            ZIDRef_Available fp s ->
-              pure $ Just (fp, s)
-      let zs = parseZettels extractQueriesWithContext $ Map.toList filesWithContent
+        flip Map.traverseMaybeWithKey zidRefs $ \zid -> \case
+          R.ZIDRef_Ambiguous fps -> do
+            tell $ one (zid, ZettelError_AmbiguousID fps)
+            pure Nothing
+          R.ZIDRef_Available fp s pluginData ->
+            pure $ Just (fp, (s, pluginData))
+      let zs = Plugin.afterZettelParse plugins (Map.toList filesWithContent)
       g <- G.buildZettelkasten zs
       pure (g, zs)
-
--- | What does a Zettel ID refer to?
-data ZIDRef
-  = -- | The ZID maps to a file on disk with the given contents
-    ZIDRef_Available FilePath Text
-  | -- | The ZID maps to more than one file, hence ambiguous.
-    ZIDRef_Ambiguous (NonEmpty FilePath)
-  deriving (Eq, Show)
-
-resolveZidRefs :: [FilePath] -> Action (Map ZettelID ZIDRef)
-resolveZidRefs files = do
-  notesDir <- ribInputDir
-  -- Use State monad to "gather" duplicate zettel files using same IDs, in the
-  -- `Right` of the Either state value; with the `Left` collecting the actual
-  -- zettel files to load into the graph.
-  fmap snd $
-    flip runStateT (Map.empty :: Map ZettelID ZIDRef) $ do
-      forM_ files $ \relPath -> do
-        whenJust (getZettelID relPath) $ \zid -> do
-          gets (Map.lookup zid) >>= \case
-            Just (ZIDRef_Available oldPath _s) -> do
-              -- The zettel ID is already used by `oldPath`. Mark it as a dup.
-              modify $ Map.insert zid (ZIDRef_Ambiguous $ relPath :| [oldPath])
-            Just (ZIDRef_Ambiguous (toList -> ambiguities)) -> do
-              -- Third or later duplicate file with the same Zettel ID
-              modify $ Map.insert zid (ZIDRef_Ambiguous $ relPath :| ambiguities)
-            Nothing -> do
-              let absPath = notesDir </> relPath
-              lift $ need [absPath]
-              s <- decodeUtf8With lenientDecode <$> readFileBS absPath
-              modify $ Map.insert zid (ZIDRef_Available relPath s)