+++ neuron/src/lib/Neuron/Web/Query/View.hs
@@ -25,7 +25,6 @@ import Data.Some (Some (..))
 import Data.TagTree
   ( Tag (..),
     TagNode (..),
-    TagPattern (..),
     constructTag,
     foldTagTree,
     tagMatchAny,
@@ -107,18 +106,14 @@ renderQuery someQ =
     case someQ of
       Some (ZettelQuery_ZettelByID _ _) ->
         blank
-      Some (ZettelQuery_ZettelsByTag [] _mconn _mview) ->
-        text "All zettels"
-      Some (ZettelQuery_ZettelsByTag (fmap unTagPattern -> pats) _mconn _mview) -> do
-        let qs = toText $ intercalate ", " pats
+      Some (ZettelQuery_ZettelsByTag q _mconn _mview) -> do
+        let qs = show q
             desc = toText $ "Zettels tagged '" <> qs <> "'"
         elAttr "span" ("class" =: "ui basic pointing below black label" <> "title" =: desc) $ do
           semanticIcon "tags"
           text qs
-      Some (ZettelQuery_Tags []) ->
-        text "All tags"
-      Some (ZettelQuery_Tags (fmap unTagPattern -> pats)) -> do
-        let qs = toText $ intercalate ", " pats
+      Some (ZettelQuery_Tags q) -> do
+        let qs = show q
         text $ "Tags matching '" <> qs <> "'"
       Some (ZettelQuery_TagZettel _tag) -> do
         blank
@@ -153,7 +148,7 @@ renderZettelLink mInner conn (fromMaybe def -> linkView) Zettel {..} = do
         -- The extra space is so that double clicking on this extra text
         -- doesn't select the title next.
         text " "
-    elAttr "span" ("class" =: "zettel-link" <> withTooltip linkTooltip) $ do
+    elAttr "span" ("class" =: "zettel-link" <> maybe mempty ("title" =:) linkTooltip) $ do
       let linkInnerHtml = fromMaybe (text zettelTitle) mInner
       neuronRouteLink (Some $ Route_Zettel zettelSlug) mempty linkInnerHtml
       elConnSuffix conn
@@ -164,14 +159,6 @@ renderZettelLink mInner conn (fromMaybe def -> linkView) Zettel {..} = do
       | isJust mInner = Just $ "Zettel: " <> zettelTitle
       | null zettelTags = Nothing
       | otherwise = Just $ "Tags: " <> T.intercalate "; " (unTag <$> zettelTags)
-    withTooltip :: Maybe Text -> Map Text Text
-    withTooltip = \case
-      Nothing -> mempty
-      Just s ->
-        ( "data-tooltip" =: s
-            <> "data-inverted" =: ""
-            <> "data-position" =: "right center"
-        )
 
 elConnSuffix :: DomBuilder t m => Maybe Connection -> m ()
 elConnSuffix mconn =