+++ neuron/src/lib/Neuron/Zettelkasten/Zettel.hs
@@ -5,6 +5,7 @@
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE GADTs #-}
 {-# LANGUAGE LambdaCase #-}
+{-# LANGUAGE MultiParamTypeClasses #-}
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE RankNTypes #-}
 {-# LANGUAGE RecordWildCards #-}
@@ -20,17 +21,20 @@ module Neuron.Zettelkasten.Zettel where
 
 import Data.Aeson
 import Data.Aeson.GADT.TH (deriveJSONGADT)
+import Data.Dependent.Map (DMap)
 import Data.Dependent.Sum.Orphans ()
-import Data.GADT.Compare.TH (DeriveGEQ (deriveGEq))
+import Data.GADT.Compare.TH
+  ( DeriveGEQ (deriveGEq),
+  )
 import Data.GADT.Show.TH (DeriveGShow (deriveGShow))
 import Data.Graph.Labelled (Vertex (..))
 import Data.Some (Some)
-import Data.TagTree (Tag, TagPattern (..))
+import Data.TagTree (Tag, TagQuery)
 import Data.Time.DateMayTime (DateMayTime)
 import Neuron.Markdown (ZettelParseError)
+import Neuron.Plugin.PluginData (PluginData)
 import Neuron.Zettelkasten.Connection (Connection)
 import Neuron.Zettelkasten.ID (Slug, ZettelID)
-import Neuron.Zettelkasten.Query.Error (QueryResultError)
 import Neuron.Zettelkasten.Query.Theme (ZettelsView)
 import Relude hiding (show)
 import Text.Pandoc.Builder (Block)
@@ -40,10 +44,13 @@ import Text.Show (Show (show))
 -- | ZettelQuery queries individual zettels.
 --
 -- It does not care about the relationship *between* those zettels; for that use `GraphQuery`.
+--
+-- NOTE: This type is defined in this module, rather than Zettel.Query, because
+-- of the mutual dependency with the `ZettelT` type.
 data ZettelQuery r where
   ZettelQuery_ZettelByID :: ZettelID -> Connection -> ZettelQuery Zettel
-  ZettelQuery_ZettelsByTag :: [TagPattern] -> Connection -> ZettelsView -> ZettelQuery [Zettel]
-  ZettelQuery_Tags :: [TagPattern] -> ZettelQuery (Map Tag Natural)
+  ZettelQuery_ZettelsByTag :: TagQuery -> Connection -> ZettelsView -> ZettelQuery [Zettel]
+  ZettelQuery_Tags :: TagQuery -> ZettelQuery (Map Tag Natural)
   ZettelQuery_TagZettel :: Tag -> ZettelQuery ()
 
 -- | A zettel note
@@ -64,24 +71,13 @@ data ZettelT content = Zettel
     -- | List of all queries in the zettel
     zettelQueries :: [(Some ZettelQuery, [Block])],
     zettelParseError :: Maybe ZettelParseError,
-    zettelContent :: content
+    zettelContent :: content,
+    zettelPluginData :: DMap PluginData Identity
   }
   deriving (Generic)
 
 newtype MetadataOnly = MetadataOnly ()
   deriving (Generic, ToJSON, FromJSON)
-
--- | All possible errors for a given zettel ID
-data ZettelError
-  = -- | The zettel file content is malformed
-    ZettelError_ParseError (Slug, ZettelParseError)
-  | -- | Some queries in zettel file are incorrect
-    ZettelError_QueryResultErrors (Slug, NonEmpty QueryResultError)
-  | -- | A zettel ID may refer one of several zettel files
-    ZettelError_AmbiguousID (NonEmpty FilePath)
-  | -- | A slug is shared more than one zettel file
-    ZettelError_AmbiguousSlug Slug
-  deriving (Eq, Show, Generic, ToJSON, FromJSON)
 
 -- | Zettel without its content
 type Zettel = ZettelT MetadataOnly